sketch for padding class
